fix(publish): modify the location of the read/write push prop method
